[Misc] Move `LRUCache` into its own file
