+++ vllm/multimodal/cache.py
@@ -17,7 +17,8 @@ from vllm.distributed.device_communicators.shm_object_storage import (
 )
 from vllm.envs import VLLM_OBJECT_STORAGE_SHM_BUFFER_NAME
 from vllm.logger import init_logger
-from vllm.utils import GiB_bytes, LRUCache, MiB_bytes
+from vllm.utils import GiB_bytes, MiB_bytes
+from vllm.utils.cache import LRUCache
 from vllm.utils.jsontree import json_count_leaves, json_map_leaves, json_reduce_leaves
 
 from .inputs import (
